backend-rest/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
mangum==0.17.0
pydantic==2.7.4
pydantic-settings==2.3.4
supabase==2.5.1
PyJWT==2.8.0
resend==2.0.0
httpx==0.27.0
pywebpush==2.0.0
python-multipart==0.0.9
email-validator==2.2.0
